requirements.txt
tweepy
pandas==1.5.3
numpy
SQLAlchemy
praw
geopy
boto3
python-dotenv
psycopg2-binary
openpyxl